# Switch to a custom Myers-based LCS engine