--- a/src/context_diff.rs
+++ b/src/context_diff.rs
@@ -616,6 +616,7 @@ mod tests {
                                 let _ = fb;
                                 let output = Command::new("patch")
                                     .arg("-p0")
+                                    .arg(format!("{target}/alefx"))
                                     .arg("--context")
                                     .stdin(File::open(format!("{target}/abx.diff")).unwrap())
                                     .output()