--- a/src/unified_diff.rs
+++ b/src/unified_diff.rs
@@ -771,6 +771,7 @@ mod tests {
                                 let _ = fb;
                                 let output = Command::new("patch")
                                     .arg("-p0")
+                                    .arg(format!("{target}/alefx"))
                                     .stdin(File::open(format!("{target}/abx.diff")).unwrap())
                                     .output()
                                     .unwrap();