This is a big one. It basically allows our version of diff to be used on larger files, which right now crash due to an impossibly large allocation, due to the library we use implementing the LCS algorithm using a full matrix matching lines of both files.

I looked at implementing this on the library, but the author already has a patch for months and has been unresponsive. I also think it makes sense to have our own engine that can be customized and optimized as required by the goals of the tool rather than relying on a library which may be more conservative. The author mentioned in their PR that the plan there is to make the Myers diagonal search optional due to the pathological cases - which to me means heuristics will probably not be welcome.

I split the actual feature in 3 commits:

- implementation and adoption of the Myers diagonal search
- importing some heuristics used by GNU diff to bail out on "good enough" or "too expensive" searches
- adapting yet another heuristic to give up on the algorithm completely for files that seem just too large and different

See each commit for more details of performance, it is looking pretty good. I would prefer to keep the commits separate, like I did for the cmp optimizations, so that we have a good base for debugging and bisecting if we find some weird corner case.

Tests all pass, the diffs produced I looked manually at seem to have good quality. To weed out any corner cases (and I did find some!) I have been running the following script for the last couple of days. Since I last fired it up it's been going for over 12 hours now, basically diffing every single text and source code file I have in my home directory (over 370k files, I have multiple browser code bases fwiw) against /etc/passwd, an empty file, the previous file, all of them back and forth, then trying to apply it with patch and checking for correctness for each of those. So far no further issues found!

```python
#!/bin/env python3
import os
import subprocess
import sys
import shutil

def run_diff(a, b):
    shutil.copyfile(a, "a");
    shutil.copyfile(b, "b");
    ret = subprocess.run(f"./target/debug/diffutils diff a b".split(' '), stdout=open('diff', 'w'))
    print(f"ret: {ret}")
    if ret.returncode < 0 or ret.returncode == 101:
        print(f"diff killed: {ret.returncode}")
        sys.exit(2)

    ret = subprocess.run(f"patch -p0 a diff".split(' '))
    print(f"ret: {ret}")
    if ret.returncode != 0:
        print(f"patch failed: {ret.returncode}")
        sys.exit(2)

    assert open('a', 'rb').read() == open('b', 'rb').read()

previous_file = '/etc/passwd'
for current_file in open('../diff.rs/text-files'):
    current_file = current_file.strip()
    print(f"Diffing {previous_file} to {current_file}")
    for reference in ['/etc/passwd', 'empty', previous_file]:
        run_diff(reference, current_file)
        run_diff(current_file, reference)
    previous_file = current_file
```