@sylvestre 2024-10-27
Interesting. 
Did you try with hyperfine for benchmarking ? Thanks

Maybe create a BENCHMARKING.md file to document the improvement ?