@kov 2024-10-27
> Interesting. Did you try with hyperfine for benchmarking ? Thanks

Yep! There are some details in the individual commit messages, but the most important improvement from my perspective is being able to handle larger files at all. For very short files it should be more or less the same.

The worst case for this algorithm is completely different files, and for those (on files short enough for the old version to be able to process ;D) we have a performance regression:

Before:

> hyperfine -N -i --output=pipe --warmup 2 './target/release/diffutils diff -u t/COPYING t/COPYING2'
Benchmark 1: ./target/release/diffutils diff -u t/COPYING t/COPYING2
  Time (mean ± σ):       3.1 ms ±   0.3 ms    [User: 1.9 ms, System: 1.1 ms]
  Range (min … max):     2.6 ms …   6.9 ms    665 runs

After:

> hyperfine -N -i --output=pipe --warmup 2 './target/release/diffutils diff -u /usr/share/licenses/glib2/LGPL-2.1-or-later.txt /usr/share/licenses/coreutils-common/COPYING'
Benchmark 1: ./target/release/diffutils diff -u /usr/share/licenses/glib2/LGPL-2.1-or-later.txt /usr/share/licenses/coreutils-common/COPYING
  Time (mean ± σ):       4.8 ms ±   0.3 ms    [User: 3.9 ms, System: 0.8 ms]
  Range (min … max):     4.2 ms …   6.8 ms    444 runs

> Maybe create a BENCHMARKING.md file to document the improvement ?

Will do later today!