@sylvestre 2024-10-27
Please run hyperfine once with the two commands. It produces better results and comparison