@oSoMoN 2024-11-01
I have yet to do a proper code review, but the approach looks sound to me, and a huge undertaking, so thank you very much for this!

@sylvestre any objection to not depending on the external diff crate and re-implementing it in-house?

@kov can you add a heading comment to `src/engine.rs` that briefly explains the algorithm being implemented, together with the heuristics, like you did in the commit messages? The GNU diff implementation also quotes and provides [a (paywall) link](http://dx.doi.org/10.1007/BF01840446) to the original paper by Eugene W. Myers, could you add that too?