@sylvestre 2024-11-02
> @sylvestre any objection to not depending on the external diff crate and re-implementing it in-house?

fine with that but I am wondering if the algo itself could not a into a different crate (in case others would like to use it) ?