@kov 2024-11-05
> > @sylvestre any objection to not depending on the external diff crate and re-implementing it in-house?
> 
> fine with that but I am wondering if the algo itself could not a into a different crate (in case others would like to use it) ?

I'd be fine with that, do you envision it being a crate of the uutils project or do yu mean external? When I pondered this as I was starting I thought it may make more sense if we exported the diffutils crate as a library because then you'd get the whole functionality, but there is a benefit to having the generic implementation that can be used for file lines or anything else exposed separately, I suppose.