@sylvestre 2024-11-08
yeah, we can create https://github.com/uutils/diff-mayers-based-lcs-engine for example :)