@sylvestre 2024-11-11
i noticed this: https://crates.io/crates/lcs-diff
did you try with it ? :)