@kov 2024-11-24
I have been a bit busy the last couple weeks with work, but I'll get back to this on the holidays and get the standalone crate going, I'll close this one meanwhile, thank you for the the input so far!